financial_analytics/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
frozendict==2.4.6
html5lib==1.1
idna==3.10
lxml==5.3.0
multitasking==0.0.11
numpy==1.21.6
pandas==1.3.5
peewee==3.17.7
platformdirs==4.0.0
py4j==0.10.9.7
pyspark==3.4.4
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.2
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
typing-extensions==4.7.1
urllib3==2.0.7
webencodings==0.5.1
yfinance==0.2.48

# Adding Snowflake dependencies
snowflake-connector-python==3.0.4
snowflake-sqlalchemy==1.5.1
pyarrow==10.0.1  # Required for Pandas integration with Snowflake